third_party/WebKit/LayoutTests/fast/dom/HTMLAnchorElement/remove-href-from-focused-anchor.html

<!DOCTYPE html>
<body>
<a href="javascript:" id="target">link</a>
<script src="../../../resources/js-test.js"></script>
<script>
jsTestIsAsync = true;
var target = document.getElementById('target');
target.focus();
shouldBe('document.activeElement', 'target');
debug('Remove href.');
window.onload = function() {
    target.addEventListener('blur', function() {
        testPassed('blur event was disaptched.');
        target.remove();
        finishJSTest();
    }, false);
    // Need to wait until CheckFocusedElementTask is unqueued.
    setTimeout(function() {
        target.removeAttribute('href');
    }, 10);
    setTimeout(function() {
        testFailed('No blur event');
        finishJSTest();
    }, 2000);
};
</script>
</body>
